storm/aclagent: rename fakes package to proxies, drop harness unit tests
Rename tools/storm/aclagent/fakes -> proxies (aks-rp and kubelet
proxies used by the storm e2e harness). Also drop
apiserver_test.go/rp_test.go: this is test-harness code exercised
end-to-end by every storm-trident aclagent run, so a broken proxy
fails the e2e scenario the same way a failing unit test would,
without the added maintenance of separate unit coverage.

Copilot-Session: 8c06585d-a82d-475f-a802-83fdfa012d86
